fix(tests): Rename integration tests package to 'integration' and use test.SkipIfCI in TestCPUTemperature
